isolation-forest/requirements.txt
asyncpg==0.31.0
joblib==1.4.2
matplotlib==3.9.2
pandas==2.2.3
python-dotenv==1.2.2
scikit-learn==1.6.1
seaborn==0.13.2
SQLAlchemy==2.0.49
